:arrow_right: Rename /configurator folder to /sidebar
