+++ src/sidebar/Configurator.js
