+++ src/sidebar/Configurator.scss
