+++ src/sidebar/Configurator.spec.js
