+++ src/sidebar/Configurator.template.html
