+++ src/sidebar/Preview.js
