+++ src/sidebar/Preview.scss
