+++ src/sidebar/options/EarthOptions.js
