+++ src/sidebar/options/FireOptions.js
